refactor(runner): split run-experiment
